Rename reducer and action for errorsMain
