+++ src/errors-store.js
@@ -1,7 +1,7 @@
 import { combineReducers } from "redux";
 import { intlReducer } from "react-intl-redux";
 import { routerReducer } from "react-router-redux";
-import errorsReducer from "./login/redux/reducers/ErrorsMain";
+import errorsReducer from "./login/redux/reducers/errorsMainReducer";
 import notificationsReducer from "reducers/Notifications";
 
 const eduIDApp = combineReducers({
