+++ src/login/redux/reducers/errorsMainReducer.js
@@ -1,4 +1,4 @@
-import * as actions from "../actions/ErrorsMain";
+import * as actions from "../actions/errorsMainActions";
 
 const configData = {
   is_app_loaded: false,
