+++ src/login/redux/sagas/ErrorsMain.js
@@ -1,6 +1,6 @@
 import { put, call } from "redux-saga/effects";
 import { checkStatus, getRequest } from "sagas/common";
-import * as actions from "../actions/ErrorsMain";
+import * as actions from "../actions/errorsMainActions";
 
 export function* requestConfig() {
   try {
